# openapi.yaml
openapi: 3.0.1
info:
  title: E-Commerce REST API
  version: 1.0.0
  description: It is a simulated e-commerce app for Codecademy project.
paths:
  /auth/register:
    post:
      summary: Register a user
      description: Register a user
      operationId: register_user
      requestBody:
        description: A new user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth_register'
      responses:
        '200':
          content:
            application/json:
              examples:
                auth:
                  value: |-
                    {
                      "id": 1,
                      "username": "test1",
                      "password": "$2b$10$DY6PQ/WoIauOdpOtdaftvuLzDgRz4tCQxNwgjmu79qD9xovZVNt7.",
                      "first_name": null,
                      "last_name": null,
                      "email": "test1@gmail.com",
                      "phone": null,
                      "created_at": "2024-02-22T08:22:12.291Z",
                      "modified_at": "2024-02-22T08:22:12.291Z"
                    }
          description: Success
        '409':
          content:
            application/json:
              examples:
                auth:
                  value: null
          description: Conflict
        '500':
          content:
            application/json:
              examples:
                auth:
                  value: null
          description: Internal Server Error
        default:
          description: Default error sample response
      tags:
        - AUTH
  /auth/login:
    post:
      summary: Login a user
      description: Login a user
      operationId: user_login
      requestBody:
        description: A user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth_login'
      responses:
        '200':
          content:
            application/json:
              examples:
                auth:
                  value: |-
                    {
                      "id": 1,
                      "username": "test1",
                      "password": "$2b$10$DY6PQ/WoIauOdpOtdaftvuLzDgRz4tCQxNwgjmu79qD9xovZVNt7.",
                      "first_name": null,
                      "last_name": null,
                      "email": "test1@gmail.com",
                      "phone": null,
                      "created_at": "2024-02-22T08:22:12.291Z",
                      "modified_at": "2024-02-22T08:22:12.291Z"
                    }
          description: Success
        '401':
          content:
            application/json:
              examples:
                auth:
                  value: null
          description: Unauthorized
        '500':
          content:
            application/json:
              examples:
                auth:
                  value: null
          description: Internal Server Error
        default:
          description: Default error sample response
      tags:
        - AUTH
...
components:
  schemas:
    Auth_register:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      xml:
        name: Auth_register
    Auth_login:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      xml:
        name: Auth_login
...